% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/featureEngineering.R
\name{calcTripDistance}
\alias{calcTripDistance}
\title{Distance calculation example}
\usage{
calcTripDistance(distFrame, type = "haversine")
}
\arguments{
\item{distFrame}{a data frame that includes dropoff_lat/long, and begintrip_lat/long.}
}
\value{
dist the haversine or euclidean distance, in meters
}
\description{
This function accepts a start lat/long and end lat/long and returns the Haversine distance.
There is an additional option to use a straight euclidean distance (not recommended).
}
\examples{
timeOfDayFnc(6)
# [1] "morning"
}

