% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/featureEngineering.R
\name{featureEngineering}
\alias{featureEngineering}
\title{Feature Generation}
\usage{
featureEngineering(tripData, truncatedData, newKmeans = NULL)
}
\arguments{
\item{tripData}{A data frame with dateTime, startLat/Long, stopLat/Long, and uid.}

\item{truncatedData}{An option to reduce the dimensionality of the return frame.}
}
\value{
featureFrame A data frame with the equivalent features calcualted
}
\description{
Take a given data frame and produce a feature vector for each unique
row.
}

