% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/featureEngineering.R
\name{findClusteredLocations}
\alias{findClusteredLocations}
\title{findClusteredLocations}
\usage{
findClusteredLocations(dataFrame, NUMCENTERS = 32)
}
\arguments{
\item{dataFrame}{a data frame with the date/time (as posixct)}
}
\value{
cluster cluster locationdocu
}
\description{
This function takes every start and end location in the Uber data set
and attempts to define k clusters (using kmeans).  The k clusters is defined
above, and defaulted to 32.
}

