% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/featureEngineering.R
\name{findDayOfWeek}
\alias{findDayOfWeek}
\title{findDayOfWeek}
\usage{
findDayOfWeek(dataFrame)
}
\arguments{
\item{dataFrame}{a data frame with the date/time (as posixct)}
}
\value{
weekday returns the "Mon", "Tues", etc day of the week for a given date.
}
\description{
This function takes a data frame and returns the 'day of the week'.
}

