% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prediction.R
\name{mlogitModel}
\alias{mlogitModel}
\title{Multinomial logistic regression}
\usage{
mlogitModel(testTrip)
}
\arguments{
\item{testTrip}{an input of the shortened feature vector}
}
\value{
cModel The model S3 object.
}
\description{
This function is another exploratory attempt at using hiearachial logistic regression.
Using the endCluster's as the variable to be predicted, it's still exploratory and
not to be used.
}

