% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prediction.R
\name{multinomialHierBayesModel}
\alias{multinomialHierBayesModel}
\title{Bayesian Hiearchial multinomial logistic regression}
\usage{
multinomialHierBayesModel(testTrip)
}
\arguments{
\item{testTrip}{an input of the shortened feature vector}
}
\value{
outMCMCs The multinomial Bayesian model chains for the estimates for beta.
}
\description{
Exploratory function.  No guarantee on code safety, included for demonstration.
This function was my top pick for being able to model the dropoff location.
It creates a list structure (one for each unique ID) as the input and predictor variable.
The output is MCMC samples for the estimates for beta.
}

