% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prediction.R
\name{naiveBayesModel}
\alias{naiveBayesModel}
\title{Naive Bayes Model}
\usage{
naiveBayesModel(testTrip)
}
\arguments{
\item{testTrip}{a data frame generated from the dataProcess functions.}
}
\value{
nbModel The entire S3 object for the NB model
}
\description{
This function takes in a data frame and returns back a naive bayes model.
}

