% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prediction.R
\name{naiveBayesTimeAnalysis}
\alias{naiveBayesTimeAnalysis}
\title{naiveBayesTimeAnalysis}
\usage{
naiveBayesTimeAnalysis(model, testData)
}
\arguments{
\item{model}{a naive bayes model}

\item{testData}{the input features from the testing data set.}
}
\value{
cfMatrix The confusion matrix from the NB analysis.
}
\description{
This function takes a model and testData, and returns a confusion matrix with
corresponding classification errors..
}

