% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{preprocessData}
\alias{preprocessData}
\title{Preprocessing of CSV data}
\usage{
preprocessData(fileInput)
}
\arguments{
\item{fileName}{A filename in string format.}
}
\value{
tripData A data frame of the pre-processed csv file.
}
\description{
Load in a dataset for training and pre-process for
acceptable data inputs
}

