% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/featureEngineering.R
\name{timeOfDayFnc}
\alias{timeOfDayFnc}
\title{Time of Day parser}
\usage{
timeOfDayFnc(tripFrame)
}
\arguments{
\item{hourVal}{A stripped out single hour.}
}
\value{
catHour A category of the
}
\description{
This function takes in an hour and parses it into a categorical variable.
}
\examples{
timeOfDayFnc(6)
# [1] "morning"
}

